Fixes #181 - Refactor modifiers to make them more maintainable
